services/ledgerexporter: Move ledgerexporter out of experimental
