+++ services/ledgerexporter/architecture.png
