+++ services/ledgerexporter/internal/test/no_network.toml
@@ -1,0 +1,11 @@
+
+[datastore_config]
+type = "ABC"
+
+[datastore_config.params]
+destination_bucket_path = "your-bucket-name/subpath"
+
+[exporter_config]
+ledgers_per_file = 3
+files_per_partition = 1
+file_suffix = ".xdr.gz"
